Gas Leakage Monitoring And Alerting System
This system provides the information such as when a gas leakage is. noticed, sensors of in the project are used to notice the gas leakage and immediately turns ON the buzzer for the danger. indication. Buzzer is a clear indication of gas leakage. By the detection of the hazardous gas the alerting message.